backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langgraph>=0.0.43
langchain-core>=0.1.46
langchain-groq>=0.1.0
pydantic>=2.5.3
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pymysql>=1.1.0